fix : 게시글 수정/삭제 권한 검사 추가

+++ src/main/java/com/sss/subscriptionsharing/web/controller/PostController.java
@@ -8,6 +8,7 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.SessionAttribute;
 
 import com.sss.subscriptionsharing.domain.Post;
 import com.sss.subscriptionsharing.domain.user.User;
+import com.sss.subscriptionsharing.exception.NoAuthorityException;
 import com.sss.subscriptionsharing.service.PostService;
 import com.sss.subscriptionsharing.service.UserService;
 import com.sss.subscriptionsharing.web.dto.PostDto;
@@ -60,15 +62,34 @@ public class PostController {
 	}
 
 	@PostMapping("/edit/post/{postId}")
-	public ResponseEntity editPost(@PathVariable Long postId, @RequestBody Map<String, String> postForm) {
+	public ResponseEntity editPost(@PathVariable Long postId, @RequestBody Map<String, String> postForm,
+		@SessionAttribute(name = LOGIN_USER, required = false) User loginUser) {
+
+		validateAuthority(postId, loginUser);
 
 		postService.edit(postId, postForm.get("title"), postForm.get("content"));
 
 		return ResponseEntity.ok().build();
 	}
 
+	private void validateAuthority(Long postId, User loginUser) {
+		Post post = postService.findById(postId).get();
+
+		if (loginUser.getId() != post.getUser().getId()) {
+			throw new NoAuthorityException("권한이 없습니다.");
+		}
+	}
+
+	@ExceptionHandler(NoAuthorityException.class)
+	public ResponseEntity noAuthority() {
+		return ResponseEntity.badRequest().build();
+	}
+
 	@PostMapping("/delete/post/{postId}")
-	public ResponseEntity deletePost(@PathVariable Long postId) {
+	public ResponseEntity deletePost(@PathVariable Long postId,
+		@SessionAttribute(name = LOGIN_USER, required = false) User loginUser) {
+
+		validateAuthority(postId, loginUser);
 
 		postService.delete(postId);
